hireblind/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pypdf==4.2.0
python-docx==1.1.2
email-validator

